moved all auth related files into auth folder
