+++ angular/src/app/auth/models/configuration.ts
