+++ angular/src/app/auth/models/oauth2-configuration.ts
