+++ angular/src/app/auth/models/oidc-user-data.ts
